Adds feature "Download as PNG"
This adds a link to download the current layout of the svg as an image in PNG
format. For browsers that support the "download" attribute on a tags, a direct
download is triggered and for the remaining browsers, a new tab is opened where
a generated image is inserted, for which right clicking and saving is an option.

Minor change to the rendered svg as well. To avoid a transparent PNG on
download, a rectangle that fills the entire svg is set as background with the
color black.