Updates the logic for handling png downloads
Instead of opening download.html with a message saying to right click and save
the generated png, a download triggered directly from filesaver.js is used
instead. Loading of canvg is still deferred to the download bundle, so the extra
100k isn't downloaded until the save png link is actually clicked.

There's still a special case for Safari, since it doesn't currently support
direct downloads like the other browsers.